+++ src/editor-ui/index.ts
@@ -1,3 +1,4 @@
 export { FoundationScreen } from './foundation/FoundationScreen'
+export { EditorShell } from './editor/EditorShell'
 export * from './primitives'
 export { colorThemes, type ColorToken, type ThemeMode } from './theme/tokens'